+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponseragas.py
@@ -61,9 +61,9 @@ class EvaluatorResponseRagas(BaseModel):
 
     model: str
 
-    created: Optional[str] = "2026-08-20T11:48:22.539Z"
+    created: Optional[str] = "2026-08-20T13:19:51.038Z"
 
-    updated: Optional[str] = "2026-08-20T11:48:22.539Z"
+    updated: Optional[str] = "2026-08-20T13:19:51.038Z"
 
     updated_by_id: OptionalNullable[str] = UNSET