+++ packages/orq-rc/src/orq_ai_sdk/models/updateannotationqueueop.py
@@ -164,7 +164,7 @@ class UpdateAnnotationQueueResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-08-20T11:48:20.273Z")
+    updated: Optional[datetime] = parse_datetime("2026-08-20T13:19:48.930Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")