+++ packages/orq-rc/RELEASES.md
@@ -10349,3 +10349,13 @@ Based on:
 - [python v4.14.0-rc.40] packages/orq-rc
 ### Releases
 - [PyPI v4.14.0-rc.40] https://pypi.org/project/orq-ai-sdk/4.14.0-rc.40 - packages/orq-rc
+
+## 2026-08-20 13:22:48
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.795.0 (2.931.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.14.0-rc.41] packages/orq-rc
+### Releases
+- [PyPI v4.14.0-rc.41] https://pypi.org/project/orq-ai-sdk/4.14.0-rc.41 - packages/orq-rc